+++ src/windows/tests.rs
@@ -5,3 +5,7 @@ mod stats;
 #[cfg(test)]
 #[path = "tests/lock.rs"]
 mod lock;
+
+#[cfg(test)]
+#[path = "tests/allocation.rs"]
+mod allocation;